+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG for husky
 
+* dev-develop
+    * FEATURE     #706 Created tiles view for datagrid
+
 * 0.25.0 (2016-08-11)
     * BUGFIX      #709 Removed action icon for ghost pages in column-navigation
     * ENHANCEMENT #700 added datagrid get-records and delete-records event